Rename ClassGenerator files to CppClassGenerator
