+++ include/Blast/CppClassGenerator.hpp
